Add support for a home position button
The home button will be rigidly attached to the body of the telescope
and such that if the focuser maximum distance the focuser could
travel into the telescope would be when the home position button is
depressed due to the focuser pushing on it.

This button will serve two functions. The first is for safety. And the other
is to get an absolute 0 position reference.